% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dep_search.R
\name{AddPredictors}
\alias{AddPredictors}
\title{Adds some more annotations based on the existing variables}
\usage{
AddPredictors(mytibble)
}
\arguments{
\item{mylist}{the list to start with}
}
\description{
Adds some more annotations based on the existing variables
}
