% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{ApplyConllFilter}
\alias{ApplyConllFilter}
\title{Applies a user-defined filter to a set of sentences}
\usage{
ApplyConllFilter(sents, filter_funct, return_type = "raw")
}
\arguments{
\item{sents}{a vector of sentences}

\item{filter_funct}{a user-defined filter function}

\item{return_type}{'raw', 'matches', 'both', 'both_pretty': raw returns just a filtered vector, matches returns the actual words that matched, both returns matches and the conll input as a separate variable and both_pretty transforms the conll input to an ordinary human readable sentence}
}
\value{
depending on the return_type parameter, either a vector of sentences or a tibble of matches
}
\description{
Applies a user-defined filter to a set of sentences
}
