% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{ApplyConllFilter_df}
\alias{ApplyConllFilter_df}
\title{Applies a filter to a data frame}
\usage{
ApplyConllFilter_df(tab, funct, sent_col, return_col = "", return_all = F)
}
\arguments{
\item{funct}{a user-defined filter function}

\item{sent_col}{the name of the column with conll data}

\item{return_col}{if specified, will add a new column to the data frame (called filtered_col) with the value of the specified column in the conll input  (e.g. lemmas of the matched words)}

\item{return_all}{if return_col as been specified and this is set to TRUE, returns the whole data frame and adds the filtered_col value to every row}

\item{a}{tibble / the data frame containing a column with conll annotated data}
}
\description{
Applies a filter to a data frame
}
