% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dep_search.R
\name{CompileAllData}
\alias{CompileAllData}
\title{Just a wrapper to  combine all the data gathering functions}
\usage{
CompileAllData(mydata)
}
\arguments{
\item{mydata}{raw list of data from the json file produced by inst/python/.....py}
}
\description{
Just a wrapper to  combine all the data gathering functions
}
