% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputting.R
\name{ConllAsSentence}
\alias{ConllAsSentence}
\title{Returns a human readable representation of the sentence}
\usage{
ConllAsSentence(tab)
}
\arguments{
\item{tab}{the raw repreentation of a sentence or a tibble}
}
\value{
the sentence in a human readable form (as a string)
}
\description{
Returns a human readable representation of the sentence
}
