% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{ContainsDepRel}
\alias{ContainsDepRel}
\title{Checks if a sentence contains a case where a is governed by b}
\usage{
ContainsDepRel(tab, depw, headw)
}
\arguments{
\item{tab}{the sentence to look in, as a tibble}

\item{depw}{c(column_name, column_value, [as_regex=T/F])}

\item{headw}{c(column_name, column_value, [as_regex=T/F])}
}
\description{
Checks if a sentence contains a case where a is governed by b
}
