% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{FilterConllRows}
\alias{FilterConllRows}
\title{Filters a tibble representing a sentence and returns matches}
\usage{
FilterConllRows(tab, column_name, column_val, use_regex = F,
  is_negative = F)
}
\arguments{
\item{tab}{a conll represented sentence as a tibble OR a string representation of the sentence which will be converted to a tibble}

\item{column_name}{e.g. POS, lemma}

\item{column_val}{what the value of the column should be, e.g N, igrat'}

\item{use_regex}{if false, exact matches will be searched, if true, regexes}

\item{is_negative}{if True, will filter only rows NOT matching the condition}
}
\value{
a filtered tibble
}
\description{
Filters a tibble representing a sentence and returns matches
}
