% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dep_search.R
\name{GetBigram}
\alias{GetBigram}
\title{Searches for the verb and  its complement}
\usage{
GetBigram(tab, verb)
}
\arguments{
\item{tab}{the concordance as a data frame}
}
\description{
Searches for the verb and  its complement
}
