% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dep_search.R
\name{GetDataFromSource}
\alias{GetDataFromSource}
\title{Trying to implement a simple ability to query dependency annotated data in
r
 Gets the source data from json}
\usage{
GetDataFromSource(data_folder)
}
\description{
Trying to implement a simple ability to query dependency annotated data in
r
 Gets the source data from json
}
