% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_verb_querys.R
\name{GetFrequencyListFromSource}
\alias{GetFrequencyListFromSource}
\title{Reads the frequency list from a file and returns a tibble}
\usage{
GetFrequencyListFromSource(data_folder)
}
\description{
Reads the frequency list from a file and returns a tibble
}
