% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataformatting.R
\name{GetSentencesFromFile}
\alias{GetSentencesFromFile}
\title{A convenience function for getting a vector of sentences}
\usage{
GetSentencesFromFile(filename)
}
\arguments{
\item{filename}{path to the source file to be used}
}
\description{
A convenience function for getting a vector of sentences
}
