% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gospel_of_john_sentences}
\alias{gospel_of_john_sentences}
\title{Another sample vector of sentences}
\format{tibble}
\usage{
gospel_of_john_sentences
}
\description{
This is the whole gospel of john from the copyright-free World English Bible
}
\keyword{datasets}
