% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_verb_querys.R
\name{MakeQueryForCrawler}
\alias{MakeQueryForCrawler}
\title{Creates a list for querying the aranea corpora}
\usage{
MakeQueryForCrawler(verbs_sample, data_folder)
}
\description{
Creates a list for querying the aranea corpora
}
