% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dep_search.R
\name{ProcessAllData}
\alias{ProcessAllData}
\title{Processes each verb in order to find what we're looking for}
\usage{
ProcessAllData(thisdata)
}
\description{
Processes each verb in order to find what we're looking for
}
