% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dep_search.R
\name{ProcessThisConcordance}
\alias{ProcessThisConcordance}
\title{Processes each concordance including this verb}
\usage{
ProcessThisConcordance(concordance, verb)
}
\description{
Processes each concordance including this verb
}
