% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_verb_querys.R
\name{TakeSampleFromFlist}
\alias{TakeSampleFromFlist}
\title{Filters the manually filtered list of verbs a little more}
\usage{
TakeSampleFromFlist(verbs_filtered)
}
\description{
Filters the manually filtered list of verbs a little more
}
