% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dep_search.R
\name{TidyList}
\alias{TidyList}
\title{Converts the list of tibbles to one single tibble}
\usage{
TidyList(mylist)
}
\arguments{
\item{mylist}{the list to start with}
}
\description{
Converts the list of tibbles to one single tibble
}
