% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucl_ref}
\alias{ucl_ref}
\title{A sample data frame conaining syntactically annotated tweets}
\format{tibble}
\usage{
ucl_ref
}
\description{
The tweets are from the end of november 2018 and consist of comments to a football match containing the word "referee"
}
\keyword{datasets}
