% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varpunen_sentences}
\alias{varpunen_sentences}
\title{A sample vector of sentences}
\format{tibble}
\usage{
varpunen_sentences
}
\description{
The sentences are from a Finnish wikipedia article
}
\keyword{datasets}
