% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataformatting.R
\name{ConllAsTibble}
\alias{ConllAsTibble}
\title{Makes a tibble of the raw conll string provided. Assumes one sentence.}
\usage{
ConllAsTibble(conll, conll_cols = c("tokenid", "token", "lemma", "pos",
  "pos2", "feat", "head", "dep", "null1", "null2"))
}
\arguments{
\item{conll}{the raw string representing a sentence}

\item{conll_cols}{the structure of the conll (names of columns)}
}
\description{
Makes a tibble of the raw conll string provided. Assumes one sentence.
}
